ownership-eval: Add an ownership_* attribute evaluator WIP
The fledgling ownership-eval checker currently implements the same thing
as defaultEvalCall(), invalidating all non-const regions. It needs to be
modified to take GIR annotations into account, so that parameters are
not marked as invalidated as appropriate. I suspect the regions for
objects will be marked as invalidated, but their pointers will not
escape, for example.