build: Remove no-undefined flag
According to a comment in configure.ac we don't link against the LLVM
libraries because they're so large and the symbols are resolved anyway
when the plugin is loaded. For that to work, we have to remove the
no-undefined flag, otherwise the linker will complain that all of the
symbols in the LLVM libraries are undefined.