scripts: Always enable core checkers
I was told by a Clang static analyzer developer that running the
analyzer without the core checkers enabled is not supported, and I've
seen that it can lead to crashes.

For this we need to make a few changes to the test programs. In
gvariant-iter we need to avoid tripping a warning from a core checker,
and in gerror-api we can remove two tests for cases that a core checker
already detects.

See https://bugs.llvm.org/show_bug.cgi?id=42816