clang-plugin: Support LLVM 9.0
The main difference here is that evalCall checkers have changed to take
a CallEvent& parameter rather than a CallExpr* parameter. For version 9
we convert the CallEvent& into a CallExpr* as soon as we receive it.

This may need to be refactored in the future, as CallEvent is the more
modern API and covers syntax that CallExpr doesn't cover, but that is a
large refactor better left for when it becomes necessary or when LLVM 9
becomes the minimum version.