assertion: Handle ExprWithCleanups
This type of expression can occur in several instances, but primarily in
C++ where a temporary object is created. In is_assertion_stmt() and
_simplify_boolean_expr(), only consider the expression and ignore the
cleanups.

Adds some machinery to the build system and the test runner script so
that we can have test files in C++, for testing C++-only features.