clang-plugin: Add option to provide extra typelib paths
This adds a --typelib-path option to the plugin, which may be given more
than once. Each time the option is given, it adds a path to the list of
typelib search paths.

In order to deal with build directories given as typelib paths, we also
make sure to skip files in the typelib search path without the .typelib
extension.

See: #39